requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
bcrypt==4.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
ecdsa==0.19.0
executing==2.1.0
fastapi==0.112.2
groq==0.11.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
icecream==2.1.3
idna==3.8
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pytest==8.3.2
python-dotenv==1.0.1
python-jose==3.3.0
PyYAML==6.0.2
requests==2.32.3
responses==0.25.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.33
starlette==0.38.4
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6
